Update use of fetch to resolve-symlinks when included in modules.
Some modules need to reference a common folder in a repo like a "bash
commons", but fetch installs only a module's directory, not the whole
repo, so what to do?

This change incorporates the new --resolve-symlinks feature to be
released in fetch v0.2.1. This way, a module can include a symlink to a
common folder, and when the module is installed, those symlink files
will be copied into the folder.