bootstrap: allow user to override BIN_DIR & USER_DATA_DIR
This tiny enhancement allows a user to specify alternate locations for the
binaries and user-data. Overriding these values to a location where the user
has write privileges will allow this script to run without root permissions.

It is probably an uncommon use case, and an advanced user should quickly notice
these optional overrides when reviewing the script since they're at the very
beginning. For this reason, I did not add a CLI parameter nor documentation.